backend/requirements.txt
# app
fastapi
uvicorn
pydantic[dotenv]
email-validator
python-jose[cryptography]
passlib
bcrypt

# db
databases[postgresql]
SQLAlchemy
alembic
psycopg2-binary 